apps/backend/requirements.txt
# FastAPI Backend Dependencies
# PromptDec - Phase 1B Backend Core

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Supabase
supabase==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring (optional)
sentry-sdk[fastapi]==1.39.2

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
